Refactor main logic into modular files
