+++ mathutil.go
@@ -1,0 +1,45 @@
+package main
+
+import "math"
+
+func cosine32(a, b []float32) float32 {
+	var dot, na, nb float32
+	for i := range a {
+		af, bf := a[i], b[i]
+		dot += af * bf
+		na += af * af
+		nb += bf * bf
+	}
+	if na == 0 || nb == 0 {
+		return 0
+	}
+	return dot / (float32(math.Sqrt(float64(na))) * float32(math.Sqrt(float64(nb))))
+}
+
+func tinyBias(label string) float32 {
+	h := fnv32(label)
+	return float32(h%997) * 1e-9
+}
+
+func fnv32(s string) uint32 {
+	const (
+		offset32 = 2166136261
+		prime32  = 16777619
+	)
+	var h uint32 = offset32
+	for i := 0; i < len(s); i++ {
+		h ^= uint32(s[i])
+		h *= prime32
+	}
+	return h
+}
+
+func clamp01(x float32) float32 {
+	if x < 0 {
+		return 0
+	}
+	if x > 1 {
+		return 1
+	}
+	return x
+}
